lightnvm: precalculate controller write boundaries
Flash controllers typically define flash pages as a collection of flash
sectors of typically 4K. Moreover, flash controllers might program flash
pages across several planes. This defines the write granurality at which
flash can be programmed. This is different for each flash controller.

In order to simplify calculations, and avoid repeating them in a per-I/O
basis, this patch pre-calculates these values as part of the device
characteristics in the bring up. This will be used in future patches
that add write buffering to support controller-agnostic writes based on
the device characteristics.

Signed-off-by: Javier González <javier@cnexlabs.com>