lightnvm: add generic helper debugging functions
ppa boundary  debugging is general to all targets. Introduce (i) a
generic function that implements boundary checks on a given
ppa. This allows targets to debug when a ppa is out of bounds before
sending it to the physical device. Also, introduce (ii) a function that
prints the contents of a ppa in its generic, human-readable format.

Signed-off-by: Javier González <javier@cnexlabs.com>