lightnvm: move generic function to lightnvm.h
Targets are assumed to used the same generic ppa format, where the
address is partitioned on ch:lun:block:pg:pl:sec. Thus, make the
function in charge og transforming the ppa address from a linear format
to the generic one available to all targets.

This function will be needed be the media manager in order to do target
mapping translations when targets are divided on different physical
partitions.

Signed-off-by: Javier González <javier@cnexlabs.com>