lightnvm: initialize target sysfs on bringup
In order to register through the sysfs interface, a driver needs to know
its kobject. On a disk, this happens during when the partition
information is added (device_add_disk), which for lightnvm happens after
the target is initialized. Thus, on target initialization, the kboject
has not been created yet. This patch adds a target function to let
targets implement their own sysfs routines.

Signed-off-by: Javier González <javier@cnexlabs.com>