lightnvm: refactor identification path
LightNVM structures maintaining the geometry and capabilities of an
Open-Channel SSD device are tight to the 1.2 spec. With the inclusion of
the 2.0 spec, these structure present redundancies and bring to naming
confusions.

This patch abstracts the geometry of the device form the spec. the
device is implementing. This simplifies the code and allows to easily
include new fields as the specifications evolve.

Signed-off-by: Javier González <javier@cnexlabs.com>